Archon_Stack/archon/requirements.txt
# Core dependencies
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
logfire>=0.8.0
pydantic>=2.6.0
httpx>=0.26.0

# Database and storage
supabase>=2.3.0
postgrest>=0.13.0
psycopg2-binary>=2.9.9

# AI and ML
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-community>=0.0.20
langchain-core>=0.1.20
chromadb>=0.4.22

# Utilities
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.3
aiofiles>=23.2.1
watchdog>=3.0.0
rich>=13.7.0
typer>=0.9.0
click>=8.1.7

# Development tools
black>=24.1.0
isort>=5.13.2
mypy>=1.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0 